rename Full binary tree class - can be used for construt binary tree - general
