+++ Practice/Tree/BinaryTree.cs
@@ -1,33 +1,64 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Microsoft.VisualBasic;
 
 namespace Practice.Tree
 {
+    /// <summary>
+    /// Binary tree implementation / Full Binary tree
+    /// This implementation can be used also to construct Full Binary Tree
+    /// Full Binary Tree 
+    /// Is a special binary tree where every node has either 2 children or 0 children.
+    /// </summary>
     public class BinaryTree
     {
-        private TreeNode Root;
-        public BinaryTree(TreeNode root = null)
+        public Node Head;
+        public BinaryTree(Node head = null)
         {
-            Root = root;
+            Head = head;
         }
 
-        public void CBTInserter(TreeNode root)
+        /// <summary>
+        /// similar with  Complete binary tree but can add null node
+        /// </summary>
+        /// <param name="value"></param>
+        /// <returns></returns>
+        public Node Insert(int? value)
         {
-            
-        }
+            var newNode = new Node(value);
 
-        public int Insert(int val)
-        {
-            return 0;
-        }
+            if (Head == null)
+                return Head = newNode;
 
-        public TreeNode Get_root()
-        {
-            return Root;
+            var nodes = new Queue<Node>();
+            nodes.Enqueue(Head);
+
+            while (nodes.Count != 0)
+            {
+                var current = nodes.Peek();
+                //check children of the current node
+
+                if (current.Left == null)
+                {
+                    current.Left = newNode;
+                    break;
+                }
+
+                if (current.Right == null)
+                {
+                    current.Right = newNode;
+                    break;
+                }
+
+                nodes.Dequeue();
+                nodes.Enqueue(current.Left);
+                nodes.Enqueue(current.Right);
+            }
+
+            return Head;
         }
     }
 }
